fix: separate cardItem on different components
